feat (SandBox/Enzo/src/Script/Player/): add charge attack, and refactor code
